+++ cypress/components/image/DynamicAttributes.svelte
