+++ cypress/components/rx/rx-users-spec.js
@@ -1,6 +1,6 @@
 /// <reference types="cypress" />
-import RxUsers from '../components/RxUsers.svelte'
-import mount from 'cypress-svelte-unit-test'
+import RxUsers from './RxUsers.svelte'
+import {mount} from 'cypress-svelte-unit-test'
 
 /* eslint-env mocha */
 describe('RxJS users', () => {
