+++ cypress/components/tutorial/HelloComponent.spec.js
