+++ cypress/components/tutorial/StaticComponent.spec.js
