+++ cypress/components/tutorial/StaticComponent.svelte
