+++ cypress/components/tutorial/stores/price.js
