+++ cypress/components/tutorial/stores/price.spec.js
